fix: get_devices only checks for the first type
When testing this in Homeassistant with an XT2 Plus, i noticed, that
the get_devices() function checks the type by checking if the type
is found in the parameter generic_type. But for this to work and check
for the type ids used in XT2 etc, we need to fix this to not only check
for the first entry in the passed list.